slightly cleaned HTML in English splash screen
Changed two ‘<br/>’s to a ‘</p><p>’. This does not change the look of the text. There is still another non-semantic <br/> nearby, but I’m not sure whether to make it a <p> or delete it, since I didn’t write this copy. So I’m leaving that for now.